backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
Pillow
numpy
# paddlepaddle>=2.6.0
# paddleocr>=2.7.0.3
opencv-python
streamlit
pandas
plotly
google-generativeai
python-dotenv
